Convert update script into project.update
